Modified LDC method now seems to be working well. There could still be improvements but it largely seems to working as planned without having to put in the energymeters etc. It does still require a short line downstream of the regulator though, it seems. Validated only on networks 5/6/8 for now.

+++ ptech18/dss_stats_funcs.py
@@ -151,7 +151,7 @@ def getHcV2(V,Cdf,v0):
 def vmM(vec,Mat): # Rowwise multiplication, i.e. np.diag(vec).dot(Mat)]
     if len(vec)!=len(Mat):
         print('Warning! vector length is not the same length as the matrix row no.')
-    MatOut = np.zeros(np.shape(Mat)); i = 0
+    MatOut = np.zeros(np.shape(Mat),dtype=Mat.dtype); i = 0
     for mati in Mat:
         MatOut[i] = mati*vec[i]; i+=1
     return MatOut
@@ -159,7 +159,7 @@ def vmM(vec,Mat): # Rowwise multiplication, i.e. np.diag(vec).dot(Mat)]
 def mvM(Mat,vec): # Columnwise multiplication, i.e. Mat.dot(np.diag(vec))
     if len(vec)!=Mat.shape[1]:
         print('Warning! vector length is not the same length as the matrix col no.')
-    MatOut = np.zeros(np.shape(Mat)); i = 0
+    MatOut = np.zeros(np.shape(Mat),dtype=Mat.dtype); i = 0
     for i in range(Mat.shape[1]):
         MatOut[:,i] = Mat[:,i]*vec[i]; i+=1
     return MatOut
